+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2022-01-25
+
+### Changed
+
+- Removed reducer from FormField.vue because of limit error
 ## 2022-01-18
 
 ### Added